[identifydeadworms]: (3) Move functions to library/modules and library/functions/image_processing
