+++ src/frontend/cellprofiler/modules/identifydeadworms.py
@@ -77,11 +77,8 @@ from cellprofiler_core.preferences import get_default_colormap
 from cellprofiler_core.setting import Binary
 from cellprofiler_core.setting.subscriber import ImageSubscriber
 from cellprofiler_core.setting.text import LabelName, Integer, Float
-from centrosome.cpmorphology import all_connected_components
-from centrosome.cpmorphology import fixup_scipy_ndimage_result as fix
-from centrosome.cpmorphology import get_line_pts
-from scipy.ndimage import binary_erosion, binary_fill_holes
-from scipy.ndimage import mean as mean_of_labels
+
+from cellprofiler_library.modules._identifydeadworms import identify_dead_worms
 
 C_WORMS = "Worm"
 F_ANGLE = "Angle"
@@ -236,36 +233,21 @@ degrees.
         #
         image_set = workspace.image_set
         image = image_set.get_image(self.image_name.value, must_be_binary=True)
-        mask = image.pixel_data
-        if image.has_mask:
-            mask = mask & image.mask
+        image_mask = image.mask if image.has_mask else None
         angle_count = self.angle_count.value
-
         #
-        # We collect the i,j and angle of pairs of points that
-        # are 3-d adjacent after erosion.
+        # Perform the identification
         #
-        i_center, j_center, angular_orientation = self.get_3d_adjacent_after_erosion(mask, angle_count)
-
-        #
-        # Find connections based on distances, not adjacency
-        #
-        first, second = self.find_adjacent_by_distance(
-            i_center, 
-            j_center, 
-            angular_orientation, 
-            self.wants_automatic_distance.value, 
-            self.worm_width.value, 
-            self.worm_length.value, 
+        center_x, center_y, angles, nlabels, label_indexes, labels = identify_dead_worms(
+            image.pixel_data,
+            image_mask,
+            self.wants_automatic_distance.value,
+            self.worm_width.value,
+            self.worm_length.value,
             self.angle_count.value,
             self.space_distance.value,
             self.angular_distance.value
-            )
-            
-        #
-        # Do all connected components.
-        #
-        center_x, center_y, angles, nlabels, label_indexes, labels = self.process_all_connected_components(first, second, i_center, j_center, angular_orientation, mask)
+        )
 
         m = workspace.measurements
         assert isinstance(m, Measurements)
@@ -293,103 +275,6 @@ degrees.
             workspace.display_data.mask = mask
             workspace.display_data.labels = labels
             workspace.display_data.count = nlabels
-
-
-
-    def get_3d_adjacent_after_erosion(
-            self, 
-            mask, #: Annotated[Image2DBinary, Field(description="Input binary image")]
-            angle_count # Annotated[int, Field(description="Number of different angles at which the template will betried"), ge=1] = 32
-        ):
-        #
-        # We collect the i,j and angle of pairs of points that
-        # are 3-d adjacent after erosion.
-        #
-        # i - the i coordinate of each point found after erosion
-        # j - the j coordinate of each point found after erosion
-        # a - the angle of the structuring element for each point found
-        #
-        i = numpy.zeros(0, int)
-        j = numpy.zeros(0, int)
-        a = numpy.zeros(0, int)
-
-        ig, jg = numpy.mgrid[0 : mask.shape[0], 0 : mask.shape[1]]
-        for angle_number in range(angle_count):
-            angle = float(angle_number) * numpy.pi / float(angle_count)
-            strel = self.get_diamond(angle)
-            erosion = binary_erosion(mask, strel)
-            #
-            # Accumulate the count, i, j and angle for all foreground points
-            # in the erosion
-            #
-            this_count = numpy.sum(erosion)
-            i = numpy.hstack((i, ig[erosion]))
-            j = numpy.hstack((j, jg[erosion]))
-            a = numpy.hstack((a, numpy.ones(this_count, float) * angle))
-    
-        return i, j, a
-    
-    def process_all_connected_components(self, first, second, i_center, j_center, angular_orientation, mask):
-        #
-        # Do all connected components.
-        #
-        if len(first) > 0:
-            ij_labels = all_connected_components(first, second) + 1
-            nlabels = numpy.max(ij_labels)
-            label_indexes = numpy.arange(1, nlabels + 1)
-            #
-            # Compute the measurements
-            #
-            center_x = fix(mean_of_labels(j_center, ij_labels, label_indexes))
-            center_y = fix(mean_of_labels(i_center, ij_labels, label_indexes))
-            #
-            # The angles are wierdly complicated because of the wrap-around.
-            # You can imagine some horrible cases, like a circular patch of
-            # "worm" in which all angles are represented or a gentle "U"
-            # curve.
-            #
-            # For now, I'm going to use the following heuristic:
-            #
-            # Compute two different "angles". The angles of one go
-            # from 0 to 180 and the angles of the other go from -90 to 90.
-            # Take the variance of these from the mean and
-            # choose the representation with the lowest variance.
-            #
-            # An alternative would be to compute the variance at each possible
-            # dividing point. Another alternative would be to actually trace through
-            # the connected components - both overkill for such an inconsequential
-            # measurement I hope.
-            #
-            angles = fix(mean_of_labels(angular_orientation, ij_labels, label_indexes))
-            angular_orientation_variance = (angular_orientation - angles[ij_labels - 1]) ** 2
-
-            vangles = fix(mean_of_labels(angular_orientation_variance, ij_labels, label_indexes))
-            
-            aa = angular_orientation.copy()
-            aa[angular_orientation > numpy.pi / 2] -= numpy.pi
-            
-            aangles = fix(mean_of_labels(aa, ij_labels, label_indexes))
-            aangular_orientation_variance = (aa - aangles[ij_labels - 1]) ** 2
-            
-            vaangles = fix(mean_of_labels(aangular_orientation_variance, ij_labels, label_indexes))
-            
-            aangles[aangles < 0] += numpy.pi
-            angles[vaangles < vangles] = aangles[vaangles < vangles]
-            #
-            # Squish the labels to 2-d. The labels for overlaps are arbitrary.
-            #
-            labels = numpy.zeros(mask.shape, int)
-            labels[i_center, j_center] = ij_labels
-        else:
-            center_x = numpy.zeros(0, int)
-            center_y = numpy.zeros(0, int)
-            angles = numpy.zeros(0)
-            nlabels = 0
-            label_indexes = numpy.zeros(0, int)
-            labels = numpy.zeros(mask.shape, int)
-
-        return center_x, center_y, angles, nlabels, label_indexes, labels
-
 
     def display(self, workspace, figure):
         """Show an informative display"""
@@ -433,7 +318,7 @@ degrees.
             #
             lcolors = colors * 0.5 + 0.5  # Wash the colors out a little
             for ii in range(count):
-                diamond = self.get_diamond(angles[ii])
+                diamond = self.get_diamond(angles[ii], self.worm_width.value, self.worm_length.value)
                 hshape = ((numpy.array(diamond.shape) - 1) / 2).astype(int)
                 iii = int(i[ii])
                 jjj = int(j[ii])
@@ -458,129 +343,6 @@ degrees.
             normalize=False,
             sharexy=plot00,
         )
-
-    def get_diamond(self, angle):
-        """Get a diamond-shaped structuring element
-
-        angle - angle at which to tilt the diamond
-
-        returns a binary array that can be used as a footprint for
-        the erosion
-        """
-        worm_width = self.worm_width.value
-        worm_length = self.worm_length.value
-        #
-        # The shape:
-        #
-        #                   + x1,y1
-        #
-        # x0,y0 +                          + x2, y2
-        #
-        #                   + x3,y3
-        #
-        x0 = int(numpy.sin(angle) * worm_length / 2)
-        x1 = int(numpy.cos(angle) * worm_width / 2)
-        x2 = -x0
-        x3 = -x1
-        y2 = int(numpy.cos(angle) * worm_length / 2)
-        y1 = int(numpy.sin(angle) * worm_width / 2)
-        y0 = -y2
-        y3 = -y1
-        xmax = numpy.max(numpy.abs([x0, x1, x2, x3]))
-        ymax = numpy.max(numpy.abs([y0, y1, y2, y3]))
-        strel = numpy.zeros((ymax * 2 + 1, xmax * 2 + 1), bool)
-        index, count, i, j = get_line_pts(
-            numpy.array([y0, y1, y2, y3]) + ymax,
-            numpy.array([x0, x1, x2, x3]) + xmax,
-            numpy.array([y1, y2, y3, y0]) + ymax,
-            numpy.array([x1, x2, x3, x0]) + xmax,
-        )
-        strel[i, j] = True
-        strel = binary_fill_holes(strel)
-        return strel
-
-    def find_adjacent_by_distance(
-            self, 
-            i_center, 
-            j_center, 
-            angular_orientation,
-            wants_automatic_distance = True,
-            worm_width = None, # Optional[int] default 100 min 1 description="This is the width (the short axis), measured in pixels, of the diamond used as a template when matching against the worm. It should be less than the width of a worm."
-            worm_length = None, # Optional[int] default 10 min 1 description="This is the length (the long axis), measured in pixels, of the diamond used as a template when matching against the worm. It should be less than the length of a worm"
-            angle_count = None, # Optinal[int] default 32 min 1 description="This is the number of different angles at which the template will be tried. For instance, if there are 12 angles, the template will be rotated by 0°, 15°, 30°, 45° … 165°"
-            space_distance = None, # Optional[float] default 5 min 1 description="Used only if not automatically calculating distance parameters Enter the distance for calculating the worm centers, in units of pixels. The worm centers must be at least many pixels apart for the centers to be considered two separate worms."
-            angular_distance = None, # Optional[float] default 30 min 1 description="Used only if automatically calculating distance parameters IdentifyDeadWorms calculates the worm centers at different angles. Two worm centers are considered to represent different worms if their angular distance is larger than this number. The number is measured in degrees."
-        ):
-        """Return pairs of worm centers that are deemed adjacent by distance
-
-        i - i-centers of worms
-        j - j-centers of worms
-        a - angular orientation of worms
-
-        Returns two vectors giving the indices of the first and second
-        centers that are connected.
-        """
-        if len(i_center) < 2:
-            return numpy.zeros(len(i_center), int), numpy.zeros(len(i_center), int)
-        if wants_automatic_distance:
-            assert worm_width is not None and worm_length is not None, "worm_width and worm_length must be provided if wants_automatic_distance is True"
-            space_distance = worm_width
-            angle_distance = numpy.arctan2(
-                worm_width, worm_length
-            )
-            angle_distance += numpy.pi / angle_count
-        else:
-            assert space_distance is not None and angular_distance is not None, "space_distance and angular_distance must be provided if wants_automatic_distance is False"
-            space_distance = space_distance
-            angle_distance = angular_distance * numpy.pi / 180
-        #
-        # Sort by i and break the sorted vector into chunks where
-        # consecutive locations are separated by more than space_distance
-        #
-        order = numpy.lexsort((angular_orientation, j_center, i_center))
-        i_center = i_center[order]
-        j_center = j_center[order]
-        angular_orientation = angular_orientation[order]
-        breakpoint = numpy.hstack(([False], i_center[1:] - i_center[:-1] > space_distance))
-        if numpy.all(~breakpoint):
-            # No easy win - cross all with all
-            first, second = numpy.mgrid[0 : len(i_center), 0 : len(i_center)]
-        else:
-            # The segment that each belongs to
-            segment_number = numpy.cumsum(breakpoint)
-            # The number of elements in each segment
-            member_count = numpy.bincount(segment_number)
-            # The index of the first element in the segment
-            member_idx = numpy.hstack(([0], numpy.cumsum(member_count[:-1])))
-            # The index of the first element, for every element in the segment
-            segment_start = member_idx[segment_number]
-            #
-            # Develop the cross-products for each segment. Each segment has
-            # member_count * member_count crosses.
-            #
-            # # of (first,second) pairs in each segment
-            cross_size = member_count ** 2
-            # Index in final array of first element of each segment
-            segment_idx = numpy.cumsum(cross_size)
-            # relative location of first "first"
-            first_start_idx = numpy.cumsum(member_count[segment_number[:-1]])
-            first = numpy.zeros(segment_idx[-1], int)
-            first[first_start_idx] = 1
-            # The "firsts" array
-            first = numpy.cumsum(first)
-            first_start_idx = numpy.hstack(([0], first_start_idx))
-            second = (
-                numpy.arange(len(first)) - first_start_idx[first] + segment_start[first]
-            )
-        mask = (
-            numpy.abs((i_center[first] - i_center[second]) ** 2 + (j_center[first] - j_center[second]) ** 2)
-            <= space_distance ** 2
-        ) & (
-            (numpy.abs(angular_orientation[first] - angular_orientation[second]) <= angle_distance)
-            | (angular_orientation[first] + numpy.pi - angular_orientation[second] <= angle_distance)
-            | (angular_orientation[second] + numpy.pi - angular_orientation[first] <= angle_distance)
-        )
-        return order[first[mask]], order[second[mask]]
 
     def get_measurement_columns(self, pipeline):
         """Return column definitions for measurements made by this module"""
