+++ src/subpackages/library/cellprofiler_library/opts/identifydeadworms.py
@@ -1,0 +1,2 @@
+# The IdentifyDeadWorms module does not have any options
+# This file is a placeholder in case options are added in the future
