+++ tests/frontend/modules/test_identifydeadworms.py
@@ -14,6 +14,7 @@ import cellprofiler.modules.identifydeadworms
 import cellprofiler_core.object
 import cellprofiler_core.pipeline
 import cellprofiler_core.workspace
+from cellprofiler_library.modules._identifydeadworms import find_adjacent_by_distance 
 
 IMAGE_NAME = "myimage"
 OBJECTS_NAME = "myobjects"
@@ -259,7 +260,7 @@ def test_find_adjacent_by_distance_empty():
     workspace, module = make_workspace(numpy.zeros((20, 10), bool))
     assert isinstance(module, cellprofiler.modules.identifydeadworms.IdentifyDeadWorms)
 
-    first, second = module.find_adjacent_by_distance(
+    first, second = find_adjacent_by_distance(
         numpy.zeros(0), numpy.zeros(0), numpy.zeros(0), module.wants_automatic_distance.value, module.worm_width.value, module.worm_length.value, module.angle_count.value, module.space_distance.value, module.angular_distance.value
     )
     assert len(first) == 0
@@ -270,7 +271,7 @@ def test_find_adjacent_by_distance_one():
     workspace, module = make_workspace(numpy.zeros((20, 10), bool))
     assert isinstance(module, cellprofiler.modules.identifydeadworms.IdentifyDeadWorms)
 
-    first, second = module.find_adjacent_by_distance(
+    first, second = find_adjacent_by_distance(
         numpy.zeros(1), numpy.zeros(1), numpy.zeros(1), module.wants_automatic_distance.value, module.worm_width.value, module.worm_length.value, module.angle_count.value, module.space_distance.value, module.angular_distance.value
     )
     assert len(first) == 1
@@ -295,7 +296,7 @@ def test_find_adjacent_by_distance_easy():
     # Break into two groups: 0-4 (5x5) and 5-9 (5x5)
     j[5:] += 10
     a = numpy.zeros(10)
-    first, second = module.find_adjacent_by_distance(i, j, a, module.wants_automatic_distance.value, module.worm_width.value, module.worm_length.value, module.angle_count.value, module.space_distance.value, module.angular_distance.value)
+    first, second = find_adjacent_by_distance(i, j, a, module.wants_automatic_distance.value, module.worm_width.value, module.worm_length.value, module.angle_count.value, module.space_distance.value, module.angular_distance.value)
     order = numpy.lexsort((second, first))
     first = first[order]
     second = second[order]
@@ -347,7 +348,7 @@ def test_find_adjacent_by_distance_hard():
         #
         unscramble = numpy.zeros(13, int)
         unscramble[scramble] = numpy.arange(13)
-        first, second = module.find_adjacent_by_distance(i, j, a, module.wants_automatic_distance.value, module.worm_width.value, module.worm_length.value, module.angle_count.value, module.space_distance.value, module.angular_distance.value)
+        first, second = find_adjacent_by_distance(i, j, a, module.wants_automatic_distance.value, module.worm_width.value, module.worm_length.value, module.angle_count.value, module.space_distance.value, module.angular_distance.value)
         assert len(first) == 9 + 16 + 25 + 1
         assert len(second) == 9 + 16 + 25 + 1
         for f, s in zip(first, second):
